FIX for #1858494: DLGTEMPLATEEX not defined correctly
git-svn-id: https://wtl.svn.sourceforge.net/svnroot/wtl/trunk@344 4beafcb9-bf40-0410-b8f5-b741cf90ec0e

+++ wtl/include/atldlgs.h
@@ -2977,7 +2977,7 @@ struct DLGITEMTEMPLATEEX
 	short y;
 	short cx;
 	short cy;
-	WORD id;
+	DWORD id;
 };
 #pragma pack(pop)
 #endif // (_ATL_VER >= 0x800)
